+++ static/app/views/alerts/rules/metric/types.tsx
@@ -1,9 +1,13 @@
 import {t} from 'sentry/locale';
-import type {ActivationConditionType, MonitorType} from 'sentry/types/alerts';
+import type {
+  ActivationConditionType,
+  AlertRuleActivation,
+  MonitorType,
+} from 'sentry/types/alerts';
 import type {MEPAlertsQueryType} from 'sentry/views/alerts/wizard/options';
 import type {SchemaFormConfig} from 'sentry/views/settings/organizationIntegrations/sentryAppExternalForm';
 
-import type {AlertRuleActivation, Incident} from '../../types';
+import type {Incident} from '../../types';
 
 export enum AlertRuleThresholdType {
   ABOVE = 0,
@@ -84,11 +88,6 @@ export type SavedTrigger = Omit<UnsavedTrigger, 'actions'> & {
 };
 
 export type Trigger = Partial<SavedTrigger> & UnsavedTrigger;
-
-export enum ActivationCondition {
-  RELEASE_CONDITION = 0,
-  DEPLOY_CONDITION = 1,
-}
 
 // Form values for creating a new metric alert rule
 export type UnsavedMetricRule = {
